+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Embedding plot: added a "No coloring" option to the "Color by" dropdown to disable coloring.
+
 ### Changed
 
 - Header menu: show collection name in Annotations entries
+- Embedding plot coloring now supports multiple categories per sample, resolving issues with toggling visibility of categories in the legend.
 
 ### Deprecated